ipyauth/demo-api/requirements.txt
flask
flask_restful
gunicorn
requests
python-dotenv
python-jose[pycryptodome]==3
